chore: updated #2 autoencoder.py & upload utils.py
